requirements.txt
langchain
langchain-huggingface
huggingface_hub
scholarly
arxiv
pandas
biopython
requests
beautifulsoup4
fpdf
streamlit